backend/requirements.txt
stripe>=7.0
python-dotenv>=1.0
django-rest-framework-simplejwt>=5.3.1
django-cors-headers>=4.3
Pillow>=10.0
requests>=2.31
Django>=4.2,<5.1
